attainable-utility-preservation/requirements.txt
numpy==1.14.5
absl-py
matplotlib
curses;sys_platform!="win32"
windows-curses;sys_platform=="win32"